backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
rembg==2.0.50
opencv-python==4.8.1.78
numpy==1.24.3
redis==5.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0